Restructured directory & add Mockito dependency & removed sample Junit test
