--- src/main/java/Calculator.java
@@ -1,8 +1,0 @@
-/**
- * Sample class for CI Test
- */
-public class Calculator {
-    public int multiply(int a, int b) {
-        return a * b;
-    }
-}
