--- src/test/java/CalculatorTest.java
@@ -1,35 +1,0 @@
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.RepeatedTest;
-import org.junit.jupiter.api.Test;
-
-/**
- * Sample Test class for CI test
- *
- * source from https://www.vogella.com/tutorials/JUnit/article.html
- */
-public class CalculatorTest {
-    private Calculator calculator;
-
-    @BeforeEach
-    public void setUp() throws Exception {
-        calculator = new Calculator();
-    }
-
-    @Test
-    @DisplayName("Simple multiplication should work")
-    public void testMultiply() {
-        assertEquals(20, calculator.multiply(4,5),
-                "Regular multiplication should work");
-    }
-
-    @RepeatedTest(5)
-    @DisplayName("Ensure correct handling of zero")
-    public void testMultiplyWithZero() {
-        assertEquals(0, calculator.multiply(0,5), "Multiple with zero should be zero");
-        assertEquals(0, calculator.multiply(5,0), "Multiple with zero should be zero");
-    }
-}
