+++ src/test/java/com/cs5098sbext/test/correctness/StaticFieldWithValue.java
@@ -1,4 +1,4 @@
-package defectcode;
+package com.cs5098sbext.test.correctness;
 
 import org.springframework.stereotype.Component;
 import org.springframework.beans.factory.annotation.Value;
